Bugfix/Gemini Structured Output
* fix gemini structured output

* update issues templates
