+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -1,6 +1,6 @@
 name: Bug Report
 description: File a bug report to help us improve
-title: '[BUG] '
+title: ''
 labels: ['bug']
 assignees: []
 body:
